changes in all report apis

+++ Tips.Grin.Api/Entities/OpenGrinParts.cs
@@ -8,7 +8,7 @@ namespace Tips.Grin.Api.Entities
         public string? ItemNumber { get; set; }
         public string? Description { get; set; }
         [Precision(13, 3)]
-        public decimal ItemType { get; set; }
+        public int ItemType { get; set; }
         public string? UOM { get; set; }
         [Precision(18, 3)]
         public decimal Qty { get; set; }
